Renaming api to routes
